thorchain<->btc dual lp working
